еее

+++ Scatches/ArduinoNano/blink/Blink.ino
@@ -1,5 +1,6 @@
+// Arduino Nano
 
-int LED = 2;
+byte LED = 2;
   
 void setup() {
   pinMode(LED, OUTPUT);
